requirements.txt
pandas==2.0.3
numpy==1.24.3
yfinance==0.2.18
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
